requirements.txt
cmsplugin-contact==1.1.2
Django==1.11.20
django-cms==3.6.0
django-debug-toolbar==1.11
django-memcached==0.1.2
django-rosetta==0.9.3
django-sekizai==1.0.0
django-tinymce==1.5.1
mailjet-rest==1.1.1
mysqlclient==1.4.2.post1
sorl-thumbnail==12.5.0
django-filebrowser==3.9.2
django-grappelli==2.10.4
python-memcached==1.59
Pillow==6.0.0
djangocms-text-ckeditor==3.2.1
djangocms-picture==2.2.0
django-recaptcha==2.0.4
psycopg2-binary<2.8
